BubbleBobble code review

+++ 3rdParty/rapidjson/internal/diyfp.h
@@ -130,7 +130,7 @@ struct DiyFp {
     void NormalizedBoundaries(DiyFp* minus, DiyFp* plus) const {
         DiyFp pl = DiyFp((f << 1) + 1, e - 1).NormalizeBoundary();
         DiyFp mi = (f == kDpHiddenBit) ? DiyFp((f << 2) - 1, e - 2) : DiyFp((f << 1) - 1, e - 1);
-        mi.f <<= mi.e - pl.e;
+        mi.f <<= static_cast<uint64_t>(mi.e) - static_cast<uint64_t>(pl.e);
         mi.e = pl.e;
         *plus = pl;
         *minus = mi;
