+++ 3rdParty/rapidjson/writer.h
@@ -357,8 +357,8 @@ protected:
         };
 
         if (TargetEncoding::supportUnicode)
-            PutReserve(*os_, 2 + length * 6); // "\uxxxx..."
-        else
+            PutReserve(*os_, static_cast<size_t>(2) + length * 6);
+    	else
             PutReserve(*os_, 2 + length * 12);  // "\uxxxx\uyyyy..."
 
         PutUnsafe(*os_, '\"');
